refactor(mui): one file per component, matching mantine
chrome.tsx and tables.tsx split into Toolbar, ActiveFilterChips,
BulkActionBar, PaginationFooter, ErrorState, TableSkeleton, FilterDrawer,
ExpandToggle, RowActionButtons, DesktopTable and MobileCards. No behaviour,
exports or public API change.
