+++ packages/adapter-mui/src/components/DesktopTable.tsx
@@ -1,3 +1,4 @@
+/** The desktop `<table>`: header, pinned columns, rows and summary. */
 import {
   type ColumnDef,
   columnResizeHandleProps,
@@ -7,8 +8,6 @@ import {
   PIN_Z,
   pinnedCellStyle,
   type RowAction,
-  runRowAction,
-  type TableLabels,
   tableMinWidth,
   useHorizontalOverflow,
 } from "@adapttable/core";
@@ -16,18 +15,18 @@ import {
   headerGroupRow,
   type PinLeads,
   pinnedColumnWidth,
-  resolveDisabledReason,
-  resolveVirtualRows,
   rowClickProps,
   rowEditingSignature,
   type SharedTableRenderProps,
   tableRenderModel,
   useSummaryCells,
 } from "@adapttable/core/adapter";
-import type { ReactElement, ReactNode } from "react";
+import type { ReactNode } from "react";
 import { memo, useCallback, useMemo, useRef } from "react";
 
-/** Sx for an absolutely-positioned column-resize handle. */
+import { ExpandToggle } from "./ExpandToggle";
+import { RowActionButtons } from "./RowActionButtons";
+
 const RESIZE_HANDLE_SX = {
   position: "absolute",
   insetInlineEnd: 0,
@@ -40,11 +39,7 @@ const RESIZE_HANDLE_SX = {
 } as const;
 import {
   Box,
-  Card,
-  CardContent,
   Checkbox,
-  IconButton,
-  Stack,
   type SxProps,
   Table,
   TableBody,
@@ -54,21 +49,19 @@ import {
   TableRow,
   TableSortLabel,
   type Theme,
-  Tooltip,
-  Typography,
 } from "@mui/material";
 
 import { EditableDataCell } from "./EditableCell";
-import { GroupHeaderCard, GroupHeaderRow } from "./GroupHeader";
+import { GroupHeaderRow } from "./GroupHeader";
 
 /** Map a destructive colour token to MUI's `"error"` palette, else default. */
-function muiColor(color: string | undefined): "default" | "error" {
+export function muiColor(color: string | undefined): "default" | "error" {
   return color === "danger" || color === "red" || color === "error"
     ? "error"
     : "default";
 }
 
-interface SharedProps<TRow> extends SharedTableRenderProps<TRow> {
+export interface SharedProps<TRow> extends SharedTableRenderProps<TRow> {
   size: "small" | "medium";
   /** Class applied to every mobile card (merged before `rowClassName`). */
   cardClassName?: string;
@@ -104,7 +97,7 @@ export function useStableToggle(
 }
 
 /** Inline chevron pointing at the reading end; rotates down when open. */
-function ExpandChevron({
+export function ExpandChevron({
   expanded,
   dir,
 }: Readonly<{ expanded: boolean; dir?: "ltr" | "rtl" }>) {
@@ -133,34 +126,6 @@ function ExpandChevron({
   );
 }
 
-/** The per-row expand/collapse chevron button (desktop cell + mobile card). */
-function ExpandToggle({
-  id,
-  expanded,
-  onToggle,
-  dir,
-  expandLabel,
-  collapseLabel,
-}: Readonly<{
-  id: string;
-  expanded: boolean;
-  onToggle: (id: string) => void;
-  dir?: "ltr" | "rtl";
-  expandLabel: string;
-  collapseLabel: string;
-}>) {
-  return (
-    <IconButton
-      size="small"
-      aria-expanded={expanded}
-      aria-label={expanded ? collapseLabel : expandLabel}
-      onClick={() => onToggle(id)}
-    >
-      <ExpandChevron expanded={expanded} dir={dir} />
-    </IconButton>
-  );
-}
-
 /**
  * Logical (RTL-aware) `text-align` for a column. Applied via `sx` rather
  * than MUI's physical `align` prop so `"end"` follows the writing direction
@@ -172,55 +137,6 @@ function muiAlign(
   if (align === "center") return "center";
   if (align === "end") return "end";
   return "start";
-}
-
-function RowActionButtons<TRow>({
-  row,
-  actions,
-  confirm,
-  cancelLabel,
-}: Readonly<{
-  row: TRow;
-  actions: RowAction<TRow>[];
-  confirm: ConfirmHandler;
-  cancelLabel: string;
-}>) {
-  return (
-    <Stack direction="row" spacing={0.5} sx={{ justifyContent: "flex-end" }}>
-      {actions.map((action) => {
-        if (action.isHidden?.(row)) return null;
-        const reason = resolveDisabledReason(action.disabledReason?.(row));
-        const disabled =
-          reason !== undefined || (action.isDisabled?.(row) ?? false);
-        return (
-          <Tooltip key={action.key} title={reason ?? action.label}>
-            <span>
-              <IconButton
-                size="small"
-                color={muiColor(action.color)}
-                disabled={disabled}
-                aria-label={action.label}
-                onClick={
-                  // The disabled attribute already blocks activation, so
-                  // attach the handler only when the action can run.
-                  disabled
-                    ? undefined
-                    : (e) => {
-                        e.stopPropagation();
-                        runRowAction(action, row, confirm, cancelLabel);
-                      }
-                }
-              >
-                {action.icon ?? (
-                  <Typography variant="caption">{action.label}</Typography>
-                )}
-              </IconButton>
-            </span>
-          </Tooltip>
-        );
-      })}
-    </Stack>
-  );
 }
 
 /**
@@ -869,316 +785,5 @@ export function DesktopTable<TRow>({
         )}
       </Table>
     </Box>
-  );
-}
-
-function mobileLabel<TRow>(column: ColumnDef<TRow>): string {
-  return (
-    column.mobileLabel ??
-    (typeof column.header === "string" ? column.header : column.key)
-  );
-}
-
-/** Per-card inputs for the memoized {@link MobileCardBase}. */
-interface MobileCardProps<TRow> {
-  row: TRow;
-  index: number;
-  /** Stable row id (selection / expansion key). */
-  id: string;
-  columns: ColumnDef<TRow>[];
-  labels: Required<TableLabels>;
-  confirm: ConfirmHandler;
-  rowActions?: RowAction<TRow>[];
-  /** Resolved `rowClassName(row, index)`, compared as a plain string. */
-  className?: string;
-  selected: boolean;
-  expanded: boolean;
-  /** Selection toggle — present only when selection is enabled. */
-  onToggleSelect?: (id: string) => void;
-  /** Expansion toggle — present only when `renderRowDetail` is set. */
-  onToggleExpand?: (id: string) => void;
-  renderDetail?: (row: TRow) => ReactNode;
-  onRowClick?: (row: TRow) => void;
-  measureElement?: (node: Element | null) => void;
-  compact: boolean;
-  dir?: "ltr" | "rtl";
-  /**
-   * Opt-in editing bundle — uncompared. Its identity changes on every
-   * keystroke anywhere in the table; the per-row visual churn is
-   * fingerprinted by `editingSignature` instead. A held card keeps an
-   * older bundle safely: its handlers read live state through refs.
-   */
-  editing?: EditableCellEditing<TRow>;
-  /** Page rows for Tab advance — uncompared (see `editing`). */
-  rows: readonly TRow[];
-  getRowId: (row: TRow) => string;
-  /** Memo digest from {@link rowEditingSignature}. */
-  editingSignature: string | null;
-}
-
-/** The card props the memo comparator deliberately skips (see `editing`). */
-type UncomparedCardProp = "editing" | "rows" | "getRowId";
-
-/** Every card prop the memo comparator checks with `Object.is`. */
-const COMPARED_CARD_PROPS: readonly Exclude<
-  keyof MobileCardProps<unknown>,
-  UncomparedCardProp
->[] = [
-  "row",
-  "index",
-  "id",
-  "columns",
-  "labels",
-  "confirm",
-  "rowActions",
-  "className",
-  "selected",
-  "expanded",
-  "onToggleSelect",
-  "onToggleExpand",
-  "renderDetail",
-  "onRowClick",
-  "measureElement",
-  "compact",
-  "dir",
-  "editingSignature",
-];
-
-/**
- * `React.memo` comparator: re-render a card only when one of its VISUAL
- * inputs changes — a search keystroke or another card's checkbox re-renders
- * the list shell, but every unchanged card bails out here.
- */
-function mobileCardPropsEqual<TRow>(
-  prev: Readonly<MobileCardProps<TRow>>,
-  next: Readonly<MobileCardProps<TRow>>
-): boolean {
-  return COMPARED_CARD_PROPS.every((key) => Object.is(prev[key], next[key]));
-}
-
-/** One card. Memoized by {@link mobileCardPropsEqual} at the call site. */
-function MobileCardBase<TRow>({
-  row,
-  index,
-  id,
-  columns,
-  labels,
-  confirm,
-  rowActions,
-  className,
-  selected,
-  expanded,
-  onToggleSelect,
-  onToggleExpand,
-  renderDetail,
-  onRowClick,
-  measureElement,
-  compact,
-  dir,
-  editing,
-  rows,
-  getRowId,
-}: Readonly<MobileCardProps<TRow>>) {
-  return (
-    <Card
-      ref={measureElement}
-      data-index={index}
-      data-stagger=""
-      variant="outlined"
-      role="listitem"
-      className={className}
-      {...rowClickProps(row, onRowClick, index)}
-    >
-      <CardContent
-        sx={compact ? { p: 1.25, "&:last-child": { pb: 1.25 } } : undefined}
-      >
-        {onToggleSelect && (
-          <Checkbox
-            slotProps={{ input: { "aria-label": labels.selectRow } }}
-            checked={selected}
-            onChange={() => onToggleSelect(id)}
-          />
-        )}
-        {onToggleExpand && (
-          <ExpandToggle
-            id={id}
-            expanded={expanded}
-            onToggle={onToggleExpand}
-            dir={dir}
-            expandLabel={labels.expandRow}
-            collapseLabel={labels.collapseRow}
-          />
-        )}
-        {columns.map((column) => (
-          <Box key={column.key} sx={{ mb: compact ? 0.5 : 1 }}>
-            <Typography
-              variant="caption"
-              color="text.secondary"
-              sx={{ display: "block" }}
-            >
-              {mobileLabel(column)}
-            </Typography>
-            {/* Cells are arbitrary ReactNode (often block elements) —
-                a <p> wrapper would be invalid HTML. */}
-            <Typography component="div" variant="body2">
-              <EditableDataCell
-                editing={editing}
-                row={row}
-                column={column}
-                rowId={id}
-                rowIndex={index}
-                rows={rows}
-                columns={columns}
-                rowKey={getRowId}
-                editLabel={labels.editCell}
-              />
-            </Typography>
-          </Box>
-        ))}
-        {rowActions && rowActions.length > 0 && (
-          <RowActionButtons
-            row={row}
-            actions={rowActions}
-            confirm={confirm}
-            cancelLabel={labels.cancel}
-          />
-        )}
-        {expanded && renderDetail && (
-          // Inside the card — and therefore inside the measured element —
-          // so virtualization keeps accurate card heights.
-          <Box sx={{ mt: 1 }}>{renderDetail(row)}</Box>
-        )}
-      </CardContent>
-    </Card>
-  );
-}
-
-/** Mobile MUI card list. */
-export function MobileCards<TRow>({
-  table,
-  cardClassName,
-  rows,
-  rowActions,
-  confirm,
-  getRowId,
-  size,
-  dir,
-  onRowClick,
-  rowClassName,
-  renderRowDetail,
-  summaryRow,
-  expansion,
-  editing,
-  grouping,
-  rowEntries,
-  paddingTop = 0,
-  paddingBottom = 0,
-  measureElement,
-}: Readonly<SharedProps<TRow>>) {
-  const { columns, selection, labels } = table;
-  const entries = resolveVirtualRows(rows, getRowId, rowEntries);
-  const compact = size === "small";
-  // Expansion is active only when BOTH halves arrived (the chrome supplies
-  // `expansion` exactly when `renderRowDetail` is set).
-  const expand = expansion && renderRowDetail ? expansion : undefined;
-  // The summary renders as a final card. Header groups and multi-sort are
-  // desktop-only concerns: cards have no column grid for a group to span and
-  // no clickable headers to shift-click.
-  const summaryCells = useSummaryCells(summaryRow, rows);
-
-  // `memo` erases generics at module level, so the memoized card is
-  // instantiated here (once — the identity is stable for the list's life).
-  const CardItem = useMemo(
-    () => memo(MobileCardBase<TRow>, mobileCardPropsEqual),
-    []
-  );
-
-  const renderCard = (row: TRow, index: number, key: string): ReactElement => {
-    const id = getRowId(row);
-    return (
-      <CardItem
-        key={key}
-        row={row}
-        index={index}
-        id={id}
-        columns={columns}
-        labels={labels}
-        confirm={confirm}
-        rowActions={rowActions}
-        className={
-          [cardClassName, rowClassName?.(row, index)]
-            .filter(Boolean)
-            .join(" ") || undefined
-        }
-        selected={selection ? selection.isSelected(id) : false}
-        expanded={expand ? expand.isExpanded(id) : false}
-        onToggleSelect={selection ? selection.toggle : undefined}
-        onToggleExpand={expand ? expand.toggle : undefined}
-        renderDetail={renderRowDetail}
-        onRowClick={onRowClick}
-        measureElement={measureElement}
-        compact={compact}
-        dir={dir}
-        editing={editing}
-        rows={rows}
-        getRowId={getRowId}
-        editingSignature={rowEditingSignature(editing, id)}
-      />
-    );
-  };
-
-  return (
-    <Stack
-      spacing={compact ? 1 : 1.5}
-      role="list"
-      aria-label={table.getTableProps()["aria-label"]}
-    >
-      {paddingTop > 0 && <Box aria-hidden sx={{ height: paddingTop }} />}
-      {grouping
-        ? grouping.entries.map((entry) =>
-            entry.kind === "group" ? (
-              <GroupHeaderCard
-                key={entry.key}
-                entry={entry}
-                selection={selection}
-                labels={labels}
-                compact={compact}
-                onToggleCollapse={(key) => grouping.collapsed.toggle(key)}
-              />
-            ) : (
-              renderCard(entry.row, entry.index, entry.key)
-            )
-          )
-        : entries.map(({ row, index, key }) => renderCard(row, index, key))}
-      {summaryCells && (
-        <Card variant="outlined" role="listitem">
-          <CardContent
-            sx={compact ? { p: 1.25, "&:last-child": { pb: 1.25 } } : undefined}
-          >
-            {columns.map((column) => {
-              const value = summaryCells[column.key];
-              // Unlike the desktop footer, a card has no columns to keep
-              // aligned, so columns without a summary are simply skipped.
-              if (value === undefined) return null;
-              return (
-                <Box key={column.key} sx={{ mb: compact ? 0.5 : 1 }}>
-                  <Typography
-                    variant="caption"
-                    color="text.secondary"
-                    sx={{ display: "block" }}
-                  >
-                    {mobileLabel(column)}
-                  </Typography>
-                  <Typography component="div" variant="body2">
-                    {value}
-                  </Typography>
-                </Box>
-              );
-            })}
-          </CardContent>
-        </Card>
-      )}
-      {paddingBottom > 0 && <Box aria-hidden sx={{ height: paddingBottom }} />}
-    </Stack>
   );
 }
