+++ packages/adapter-mui/src/components/ErrorState.tsx
@@ -1,0 +1,29 @@
+/** Load failure, with a retry. */
+import { type TableLabels } from "@adapttable/core";
+import { Alert, Button } from "@mui/material";
+
+/** Error alert. */
+export function ErrorState({
+  error,
+  labels,
+  onRetry,
+}: Readonly<{
+  error: Error;
+  labels: Required<TableLabels>;
+  onRetry?: () => void;
+}>) {
+  return (
+    <Alert
+      severity="error"
+      action={
+        onRetry ? (
+          <Button color="inherit" size="small" onClick={onRetry}>
+            {labels.retry}
+          </Button>
+        ) : undefined
+      }
+    >
+      <strong>{labels.errorTitle}</strong> — {error.message}
+    </Alert>
+  );
+}
